+++ backend/src/apiserver/resource/BUILD.bazel
@@ -18,12 +18,11 @@ go_library(
         "//backend/src/apiserver/list:go_default_library",
         "//backend/src/apiserver/metadata:go_default_library",
         "//backend/src/apiserver/model:go_default_library",
+        "//backend/src/apiserver/server:go_default_library",
         "//backend/src/apiserver/storage:go_default_library",
         "//backend/src/common/util:go_default_library",
         "//backend/src/crd/pkg/apis/scheduledworkflow/v1beta1:go_default_library",
         "//backend/src/crd/pkg/client/clientset/versioned/typed/scheduledworkflow/v1beta1:go_default_library",
-        "@com_github_argoproj_argo//errors:go_default_library",
-        "@com_github_argoproj_argo//pkg/apis/workflow:go_default_library",
         "@com_github_argoproj_argo//pkg/apis/workflow/v1alpha1:go_default_library",
         "@com_github_argoproj_argo//pkg/client/clientset/versioned/typed/workflow/v1alpha1:go_default_library",
         "@com_github_cenkalti_backoff//:go_default_library",
