+++ backend/src/apiserver/resource/resource_manager_util_test.go
@@ -18,13 +18,11 @@ import (
 	"testing"
 	"time"
 
-	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	"github.com/golang/protobuf/ptypes/timestamp"
 	api "github.com/kubeflow/pipelines/backend/api/go_client"
 	scheduledworkflow "github.com/kubeflow/pipelines/backend/src/crd/pkg/apis/scheduledworkflow/v1beta1"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func TestToSwfCRDResourceGeneratedName_SpecialCharsAndSpace(t *testing.T) {
@@ -143,30 +141,4 @@ func TestToCrdPeriodicSchedule_NilEndTime(t *testing.T) {
 		IntervalSecond: 123,
 		StartTime:      &startTime,
 	})
-}
-
-// TestResubmitWorkflowWithOnExit ensures we do not carry over the onExit node even if successful
-func TestResubmitWorkflowWithOnExit(t *testing.T) {
-	wfName := "test-wf"
-	onExitName := wfName + ".onExit"
-	wf := wfv1.Workflow{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "test-wf",
-		},
-		Status: wfv1.WorkflowStatus{
-			Phase: wfv1.NodeFailed,
-			Nodes: map[string]wfv1.NodeStatus{},
-		},
-	}
-	onExitID := wf.NodeID(onExitName)
-	wf.Status.Nodes[onExitID] = wfv1.NodeStatus{
-		Name:  onExitName,
-		Phase: wfv1.NodeSucceeded,
-	}
-	newWF, err := formulateResubmitWorkflow(&wf)
-	assert.Nil(t, err)
-	newWFOnExitName := newWF.ObjectMeta.Name + ".onExit"
-	newWFOneExitID := newWF.NodeID(newWFOnExitName)
-	_, ok := newWF.Status.Nodes[newWFOneExitID]
-	assert.False(t, ok)
 }
